requirements.txt
attrs==21.4.0
cmd2==2.4.0
pyperclip==1.8.2
wcwidth==0.2.5
discord.py==1.7.3
spotipy==2.19.0
python-dotenv==0.19.0
gunicorn==20.1.0
